Move DropPath to components/attention_blocks.py
Start a dedicated module for the dense transformer building blocks;
components.layers re-exports the name so existing imports and saved
model configs keep working.
