+++ src/graphnet/models/components/attention_blocks.py
@@ -1,0 +1,36 @@
+"""Attention and transformer blocks used in graphnet models."""
+
+from torch.functional import Tensor
+from pytorch_lightning import LightningModule
+
+
+class DropPath(LightningModule):
+    """Drop paths (Stochastic Depth) per sample."""
+
+    def __init__(
+        self,
+        drop_prob: float = 0.0,
+    ):
+        """Construct `DropPath`.
+
+        Args:
+            drop_prob: Probability of dropping a path during training.
+                If 0.0, no paths are dropped. Defaults to None.
+        """
+        super(DropPath, self).__init__()
+        self.drop_prob = drop_prob
+
+    def forward(self, x: Tensor) -> Tensor:
+        """Forward pass."""
+        if self.drop_prob == 0.0 or not self.training:
+            return x
+        keep_prob = 1 - self.drop_prob
+        shape = (x.shape[0],) + (1,) * (x.ndim - 1)
+        random_tensor = x.new_empty(shape).bernoulli_(keep_prob)
+        if keep_prob > 0.0:
+            random_tensor.div_(keep_prob)
+        return x * random_tensor
+
+    def extra_repr(self) -> str:
+        """Return extra representation of the module."""
+        return "p={}".format(self.drop_prob)
